chore: remove MCP mode entirely
kdeps mcp had no distinct value over kdeps run. Deleted cmd/mcp.go,
pkg/mcp/, WorkflowToolDef, and all docs/README references. Moved the
path-resolution helper into serve.go which still uses it for directory
support. Updated CLAUDE.md to reflect two modes only.
