+++ docs/v2/index.md
@@ -4,7 +4,7 @@ layout: home
 hero:
   name: kdeps
   text: Build and deploy AI agents in YAML.
-  tagline: Three modes - workflow pipelines, autonomous agents, and MCP servers. Export as Docker, Kubernetes, ISO, or a single binary. Works with Ollama, OpenAI, Anthropic, and any OpenAI-compatible backend.
+  tagline: Workflow pipelines and autonomous agents in YAML. Export as Docker, Kubernetes, ISO, or a single binary. Works with Ollama, OpenAI, Anthropic, and any OpenAI-compatible backend.
   image: false
   actions:
     - theme: brand
@@ -24,9 +24,6 @@ features:
   - icon: 🤖
     title: Agent Mode
     details: Run `kdeps serve` to turn your workflow into an autonomous LLM loop. Every resource auto-registers as a callable tool. The LLM plans and executes multi-step tasks.
-  - icon: 🔌
-    title: MCP Mode
-    details: Run `kdeps mcp` to expose your resources as Model Context Protocol tools. Connect to Claude Desktop, Cursor, or any MCP-compatible host.
   - icon: 📦
     title: Self-Contained Exports
     details: Export AI workflows as a single binary, Docker image, bootable ISO, or Kubernetes manifests. Ship to production without glue code.
